generation now works

+++ core/src/Actors/Consumable.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.math.Vector2;
 
 public class Consumable extends Entity {
 
-    public Consumable(Vector2 screenPos, int size) {
+    public Consumable(Vector2 screenPos, float size) {
         super(screenPos, size);
     }
